fix test_flatten
